01-mcp-study/3_chatbot/pyproject.toml
[project]
name = "3-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.28.0",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.7",
    "streamlit>=1.47.1",
]
